Added Majority Auto Support
- Fixed All Auto pathing commands
- Added Gyro Turn command, Intake command (finished), elevator command (Work in progress)
- removed Autonomous Straight (not needed)
- Added stuff to Robot
- removed duplicates from RobotMap
- added a bunch of stuff to Drive Subsystem

+++ src/org/usfirst/frc/team1334/robot/commands/Baseline.java
@@ -24,7 +24,7 @@ public class Baseline extends CommandGroup {
         // e.g. if Command1 requires chassis, and Command2 requires arm,
         // a CommandGroup containing them would require both the chassis and the
         // arm.
-    	addSequential(new AutoDriveCommand(265.658));
+    	addSequential(new AutoDriveCommand((int) 265.658));
     	
     }
 }
