+++ src/org/usfirst/frc/team1334/robot/commands/LeftRightSwitch.java
@@ -24,11 +24,11 @@ public class LeftRightSwitch extends CommandGroup {
         // e.g. if Command1 requires chassis, and Command2 requires arm,
         // a CommandGroup containing them would require both the chassis and the
         // arm.
-    	addSequential(new AutoDriveCommand(206.82));
+    	addSequential(new AutoDriveCommand((int) 206.82));
     	addSequential(new GyroTurn(90));
     	addSequential(new AutoDriveCommand(200));
     	addSequential(new GyroTurn(90));
-    	addParallel(new AutoDriveCommand(58.30));
+    	addParallel(new AutoDriveCommand((int) 58.30));
     	addSequential(new ElevatorCommand(1));
     	addSequential(new GyroTurn(90));
     	addSequential(new IntakeCommand(0));
